+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-Dl0grZMD.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-BGs4Dhcv.js";import{_ as a}from"./mermaid.core-BHmo2ehc.js";import"./chunk-5VM5RSS4-DgaIfkET.js";import"./chunk-XXDRQBXY-BmZM74bs.js";import"./chunk-POPQ4Y6H-akbO1jgY.js";import"./chunk-F27PBJKO-DSAnu5pF.js";import"./channel-D1P1PsYa.js";import"./index-DmI1qP14.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};